refactored - initialsbadge and intialslabel into one
