+++ deltachat-ios/Controller/ChatViewController.swift
@@ -84,7 +84,7 @@ class ChatViewController: MessagesViewController {
 		if let image = chat.profileImage {
 			navigationItem.rightBarButtonItem = UIBarButtonItem(image: image, style: .done, target: self, action: #selector(chatProfilePressed))
 		} else {
-			let initialsLabel = InitialsLabel(name: chat.name, color: chat.color, size: 28)
+			let initialsLabel =  InitialsBadge(name: chat.name, color: chat.color, size: 28)
 			navigationItem.rightBarButtonItem = UIBarButtonItem(customView: initialsLabel)
 		}
